refactor(types): move SchemaObject & replace PageSchema
