+++ app/domains/pageSchemas.ts
@@ -5,7 +5,7 @@ import { beratungshilfeAntragPages } from "./beratungshilfe/formular/pages";
 import { beratungshilfeVorabcheckPages } from "./beratungshilfe/vorabcheck/pages";
 import { flowIdFromPathname, parsePathname, type FlowId } from "./flowIds";
 import { kontopfaendungWegweiserPages } from "./kontopfaendung/wegweiser/pages";
-import type { SchemaObject } from "./types";
+import type { SchemaObject } from "./userData";
 
 const pages: Partial<Record<FlowId, PagesConfig>> = {
   "/beratungshilfe/vorabcheck": beratungshilfeVorabcheckPages,
@@ -33,9 +33,7 @@ export function xStateTargetsFromPagesConfig<T extends PagesConfig>(
   }));
 }
 
-// TODO: better specify PageSchema to specify enums, strings, ...
-export type PageSchema = SchemaObject;
-export type PageConfig = { pageSchema?: PageSchema; stepId: string };
+export type PageConfig = { pageSchema?: SchemaObject; stepId: string };
 export type PagesConfig = Record<string, PageConfig>;
 
 type ExtractSchemas<T extends PagesConfig> = {
